update recording fragment to take the recording parameters as input
